add proto and taskfile

+++ auth/internal/repository/db/mappers/universal_mappers.go
@@ -9,7 +9,7 @@ import (
 
 func DomainUserToAuthInfo(user enity.DomainUser) (models.AuthInfo, error) {
 	return models.AuthInfo{
-		ID:           user.ID,
+		ID:           uint64(user.ID),
 		Login:        user.Login,
 		PasswordHash: user.PasswordHash,
 		RefreshToken: user.RefreshToken,
@@ -19,7 +19,7 @@ func DomainUserToAuthInfo(user enity.DomainUser) (models.AuthInfo, error) {
 
 func AuthInfoToDomainUser(auth models.AuthInfo) (enity.DomainUser, error) {
 	return enity.DomainUser{
-		ID:           auth.ID,
+		ID:           uint(auth.ID),
 		Login:        auth.Login,
 		PasswordHash: auth.PasswordHash,
 		RefreshToken: auth.RefreshToken,
